pyproject.toml
[tool.poetry]
name = "middleware"
version = "0.1.0"
description = ""
authors = ["LucioD93 <luciodavid93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
paramiko = "^2.11.0"
scp = "^0.14.4"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

